mnq_trading_bot/requirements.txt
python-telegram-bot[job-queue]==21.10
yfinance
pandas==3.0.2
vaderSentiment==3.3.2
python-dotenv==1.2.2
requests==2.32.3
truthbrush
langchain-google-genai
langgraph==0.2.70